Bill S9225 version original - Zellnor Myrie
BILL NUMBER: S9225 SPONSOR: MYRIE TITLE OF BILL: An act to amend the penal law, in relation to the regulation of pistol converters and convertible pistols SUMMARY OF PROVISIONS: Section 1 amends subdivision 1 of section 265.00 of the penal law, which defines "Machine-gun" and adds two new subdivisions 36 and 37 to that section to define the following terms: "Pistol converter", and "Convert- ible pistol". Section 2 amends subdivision 10 of section 265.