Bill S6996A version A - Neil D. Breslin
BILL NUMBER: S6996A SPONSOR: BRESLIN TITLE OF BILL: An act to amend the judiciary law, in relation to increasing the number of county court judges in certain jurisdictions PURPOSE OR GENERAL IDEA OF BILL: This bill will increase the number of county court judges in Schenectady County from one (1) to two (2). SUMMARY OF PROVISIONS: Section 1. Subdivision 42 of section 182 of the judiciary law is amended to change the word "one" to "two".