Bill S2744A version A - Brad Hoylman-Sigal
BILL NUMBER: S2744A SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the judiciary law, in relation to increasing the amount of allowance that trial and grand jurors are entitled to in each court of the unified court system PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to increase the entitled allowance amount that trial and grand jurors receive in each court of the unified court system. SUMMARY OF PROVISIONS: Section 1.