Bill S3170C version C - James Skoufis
BILL NUMBER: S3170C SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the domestic relations law, the family court act and the civil practice law and rules, in relation to establishing "Kyra's Law" PURPOSE OR GENERAL IDEA OF BILL: To protect children by ensuring courts promote the safety of children in child custody and visitation proceedings. SUMMARY OF PROVISIONS: Section 1 of the bill identifies the name of this legislation as "Kyra's Law.