Bill S9254 version original - Anthony H. Palumbo
BILL NUMBER: S9254 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the retirement and social security law, in relation to the mandatory age of the members of the town of Southampton police department PURPOSE OR GENERAL IDEA OF BILL: Relates to the mandatory age of the Chief of the Town of Southampton Police Department. SUMMARY OF PROVISIONS: Section 1.