Bill S9248 version original - Michelle Hinchey
BILL NUMBER: S9248 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the agriculture and markets law, in relation to the beginning farmer advisory board on agriculture PURPOSE OF THE BILL: To simplify, diversify, rename, and improve the young farmer board. SUMMARY OF PROVISIONS: Section 1 of the bill renames the young farmer advisory board to be the beginning farmer advisory board.