Bill S9256 version original - Sean Ryan
BILL NUMBER: S9256 SPONSOR: RYAN TITLE OF BILL: An act to amend the local finance law, in relation to facilitating the marketing of any issue of serial bonds or notes of the city of Buffalo issued on or before a certain date PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to extend the authorization for the private sale of bonds and notes of the city of from Buffalo June 30, 2024 to June 30, 2025. SUMMARY OF PROVISIONS: Section 1 amends section 54.