Bill S9255 version original - Kevin Thomas
BILL NUMBER: S9255 SPONSOR: THOMAS TITLE OF BILL: An act to amend chapter 401 of the laws of 2002, amending the real prop- erty tax law and the Nassau county administrative code relating to assessment and review of assessments in the county of Nassau, in relation to extending certain provisions thereof SUMMARY OF PROVISIONS: Section one extends sections 2, 4, and 7 of chapter 401 of 2002 until June 30, 2026. Section two provides that the act shall take effect immediately.