Bill S8601A version A - Monica Martinez
BILL NUMBER: S8601A SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the labor law, in relation to prohibiting fingerprinting or mandatory iris and retina scanning of employees PURPOSE: The purpose of this bill is to prohibit mandatory iris and/or retina scanning of employees to retain or gain employment. SUMMARY OF PROVISIONS: Section 1. amends labor law section 201-a to include the prohibition of mandatory iris and retina scanning of employees. Section 2. is the enacting clause.