Bill S9270 version original - Brian Kavanagh
BILL NUMBER: S9270 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the election law, in relation to the term paper ballot PURPOSE OR GENERAL IDEA OF BILL: This bill would change "paper ballot" to "affidavit ballot" in section 4-117 of the election law. SUMMARY OF PROVISIONS: Section one of the bill amends section 4-117 of the election law to change "paper ballot" to "affidavit ballot". Section two sets forth the effective date.