Bill S1726A version A - Pete Harckham
BILL NUMBER: S1726A SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the public officers law, in relation to accessing records under the freedom of information law PURPOSE: The purpose of this bill is to clarify certain provisions of the Freedom of Information Law (FOIL). SUMMARY OF PROVISIONS: Section one amends the public officers' law to make various technical and clarifying amendments. Section two states the effective date.