Bill S9276 version original - Michael Gianaris
BILL NUMBER: S9276 Revised 5/24/2024 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the criminal procedure law, in relation to allowing evidence of a defendant's prior sexual offense to be admissible in a sexual assault proceeding PURPOSE: To allow courts to admit evidence that a defendant has committed prior sexual offenses in a criminal proceeding in which a defendant is accused of a sexual offense. SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new section 60.