Bill S9283 version original - Brian Kavanagh
BILL NUMBER: S9283 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the public health law, in relation to including informa- tion on congenital heart defect births in the informational leaflets required to be given to maternity patients PURPOSE OR GENERAL IDEA OF BILL: This bill would require hospitals and birth centers to include informa- tion on Congenital Heart Defects for maternity patients and their fami- lies. SUMMARY OF PROVISIONS: Section one of the bill provides the title.