Bill S6179A version A - Nathalia Fernandez
BILL NUMBER: S6179A SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the education law, in relation to establishing Eid al-Fitr and Eid al-Adha as school holidays PURPOSE OR GENERAL IDEA OF BILL: To establish Eid al-Fitr and Eid al-Adha, two Muslim holidays, as school holidays. SUMMARY OF PROVISIONS: Sections 1 and 2 amend education law to add Eid al-Fitr and Eid al-Adha to the list of days school can not be in session. Section 3 sets the effective date.