Bill S9282 version original - Dan Stec
BILL NUMBER: S9282 SPONSOR: STEC TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to an exemption for certain property from the prohibition of alcohol sales within a certain distance of a church PURPOSE: To permit an establishment at 171 Main Street, Hudson Falls, to obtain a full liquor license. The restaurant and bar is within 200 feet of a house of worship.