Bill S9068A version A - James Tedisco
BILL NUMBER: S9068A SPONSOR: TEDISCO TITLE OF BILL: An act authorizing the village of Round Lake, in the town of Malta and the county of Saratoga, to extend the lease of certain park lands in such village PURPOSE OR GENERAL IDEA OF BILL: To extend the term of the lease of certain park lands by the village of Round Lake in the town of Malta. SUMMARY OF PROVISIONS: Section 1.