Bill S9295 version original - Jessica Scarcella-Spanton
BILL NUMBER: S9295 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to authorize the Honorable Joseph J. Maltese to receive compen- sation and service credit SUMMARY OF PROVISIONS: Section 1. States the determination of the legislature and office of court administration to compensate Joseph J. Maltese, Justice for the Appellate Division of the Second Department of the Supreme Court, for his unnecessary dismissal serving the people of the State of New York. Section 2.