Bill S9292A version original - Shelley Mayer
BILL NUMBER: S9292 SPONSOR: MAYER TITLE OF BILL: An act to amend the local finance law, in relation to the period of probable usefulness of zero-emission school buses PURPOSE OR GENERAL IDEA OF BILL: To change the amortization period of electric school buses from twelve years to eight years. SUMMARY OF PROVISIONS: Section 1: amends Local Finance Law § 11(29a) to change the amortization period for the purchase of an EV school bus from 12 years to 8 years.