Bill S9297 version original - James Tedisco
BILL NUMBER: S9297 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the penal law, in relation to including aggravated ceme- tery desecration in the second degree and aggravated cemetery dese- cration in the first degree as specified offenses for purposes of hate crimes PURPOSE OR GENERAL IDEA OF BILL: To include aggravated cemetery desecration in the second degree and first degree as specified offenses for purposes of hate crime. SUMMARY OF PROVISIONS: Section 1. Subdivision 3 of section 485.