Bill S4924A version A - Jessica Ramos
BILL NUMBER: S4924a SPONSOR: RAMOS TITLE OF BILL: An act to amend the social services law, in relation to prohibiting requiring parents or caretakers to earn a minimum wage to be eligible for child care assistance SUMMARY OF PROVISIONS: Section 1. Amends subdivision 3 of section 410-w to allow social service districts to disallow rules or regulations by the Office of Children and Family Services to have a minimum earnings requirement to receive child care assistance. Section 2.