Bill S9338 version original - Jake Ashby
BILL NUMBER: S9338 SPONSOR: ASHBY TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "CW2 Casey N. Frankoski Memorial Bridge" PURPOSE: : To name a bridge in the city of Rensselaer after CW2 Casey N. Franko- ski. SUMMARY OF PROVISIONS: Section 1 amends the highway law to rename the bridge, located on Wash- ington Avenue (bridge number 1092600) which crosses over 1-90, to the "CW2 Casey N. Frankoski Memorial Bridge.