Bill S9329 version original - Pete Harckham
BILL NUMBER: S9329 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to exempting school buses from central business district tolls PURPOSE OR GENERAL IDEA OF BILL: To provide a central business district toll exemption for school buses. SUMMARY OF PROVISIONS: Section 1 of the bill would add a new subdivision 2-a to § 1704-a of the Vehicle and Traffic Law to provide an exemption from the central busi- ness district toll for school buses.