Bill S7598B version B - Rachel May
BILL NUMBER: S7598B SPONSOR: MAY TITLE OF BILL: An act to amend the public authorities law, in relation to the board of directors of the Central New York Regional Market Authority PURPOSE OR GENERAL IDEA OF BILL: To improve the governance and transparency to the public of the Central New York Regional Market Authority. SUMMARY OF PROVISIONS: Section 1: Amends Section 827 of the Public Authorities Law, amending the composition of the Board of Directors.