Bill S9337 version original - Monica Martinez
BILL NUMBER: S9337 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the education law, in relation to establishing the tran- sitional pathway program for students in higher education with autism spectrum disorder PURPOSE: The purpose of this legislation is to encourage CUNY and SUNY schools to establish transitional pathway programs to support the enrollment and matriculation of students with autism spectrum disorder. SUMMARY OF PROVISIONS: Section 1.