Bill S9356 version original - Cordell Cleare
BILL NUMBER: S9356 SPONSOR: CLEARE TITLE OF BILL: An act to amend the elder law, in relation to social model adult day services programs PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to enhance the State Office for the Aging's (SOFA) oversight capacity of social adult day care programs. SUMMARY OF PROVISIONS: Section one specifies that the rules and regulations promulgated by the Director of SOFA apply to all social adult day care programs, regardless of their source of funding.