Bill S7195B version B - Brian Kavanagh
BILL NUMBER: S7195B SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the environmental conservation law, in relation to the availability of technical assistance grants in brownfield site remedial programs PURPOSE OF BILL: This bill would allow communities to apply and receive a Technical Assistance Grant as soon as a project is approved to the Brownfield Cleanup Program. The bill also changes several references of "citizen" to "community" throughout the act.