Bill S8324A version A - Monica Martinez
BILL NUMBER: S8324A SPONSOR: MARTINEZ TITLE OF BILL: An act in relation to authorizing Calvary Tabernacle Church of God to file an application for exemption from real property taxes for certain assessment rolls PURPOSE: The purpose of this bill is to allow the town of Islip to accept an application for a retro-active property tax exemption from the Calvary Tabernacle Church of God. SUMMARY OF PROVISIONS: Section 1.