Bill S9346 version original - Brian Kavanagh
BILL NUMBER: S9346 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the general city law, the town law, and the village law, in relation to requiring each town, village, and city to develop a hous- ing action plan for everyone PURPOSE: This bill would require each city, town and village to develop a Housing Action Plan for Everyone to identify a locality's housing needs. SUMMARY OF SPECIFIC PROVISIONS: Section 1.