Bill S9374 version original - Andrew Gounardes
BILL NUMBER: S9374 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the tax law, in relation to sales of rental vehicles PURPOSE OR GENERAL IDEA OF BILL: To subject purchases of vehicles for rental car company fleets to the state sales tax. SUMMARY OF PROVISIONS: Section one of this bill clarifies that purchases of vehicles for rental company fleets shall not be considered a sale for resale for purposes of exemption from the 4% state sales tax created in Section 1105 of Tax Law.