Bill S9381 version original - Kristen Gonzalez
BILL NUMBER: S9381 SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the general business law, in relation to liability for false information provided by a chatbot PURPOSE: This bill would amend the general business law, to assign liability for the actions of chatbots to the proprietors of such chatbots. SUMMARY OF PROVISIONS: Section 1. Amends the general business taw by adding a new section 390-f, which sets forth requirements for liability of the proprietors of chatbots.