Bill S7703B version B - Brad Hoylman-Sigal
BILL NUMBER: S7703B SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring limited use motorcycles to be registered at point of sale PURPOSE OF BILL: To ensure that all limited use motorcycles are registered at the point of sale. SUMMARY OF SPECIFIC PROVISIONS: Section one amends section 2261 of the vehicle and traffic law by adding new six, which requires that all limited use motorcycles be registered at the point of sale.