Bill S9366 version original - James Sanders Jr.
BILL NUMBER: S9366 SPONSOR: SANDERS TITLE OF BILL: An act to amend the financial services law, in relation to the applica- tion of certain provisions relating to commercial financing PURPOSE OR GENERAL IDEA OF BILL: This bill would create a safe harbor from liability for lenders who comply with financial disclosure requirements in the financial services law. SUMMARY OF PROVISIONS: Section one relates to the application of Section 813 of the Financial Services Law by creating a safe harbor.