Bill S9387 version original - James Skoufis
BILL NUMBER: S9387 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the penal law, in relation to increasing penalties for assault, obstruction and harassment of election officers PURPOSE: To deter individuals from attacking or threatening an election officer while in the performance of his or her official duties. SUMMARY OF PROVISIONS: Section 1: Amends subdivision 11 of E120.