Bill S9390 version original - Anthony H. Palumbo
BILL NUMBER: S9390 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the public health law and the education law, in relation to the administration by a school or camp of over-the-counter medication to children attending such school or camp PURPOSE: Allows schools and camps to implement a policy to administer over-the- counter medication to students and children attending such schools or camps. SUMMARY OF PROVISIONS: Section 1.