Bill S9368 version original - Jessica Ramos
BILL NUMBER: S9368 SPONSOR: RAMOS TITLE OF BILL: An act to amend chapter 831 of the laws of 1981, amending the labor law relating to fees and expenses in unemployment insurance proceedings, in relation to the effectiveness thereof PURPOSE OR GENERAL IDEA OF BILL: To extend such provisions of the Labor Law relating fees and expenses for unemployment insurance proceedings for an additional two years. Under current law, these provisions will expire on December 31, 2024.