Bill S9379 version original - Pete Harckham
BILL NUMBER: S9379 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting the application of pesticides to certain local freshwater wetlands PURPOSE OR GENERAL IDEA OF BILL: To allow local governments to adopt laws which prohibit the application of pesticides to local freshwater wetlands. SUMMARY OF PROVISIONS: Section 1 of this bill amends the environmental conservation law by adding a new section 24-0513.