Bill S1050B version B - Jeremy Cooney
BILL NUMBER: S1050B SPONSOR: COONEY TITLE OF BILL: An act to amend the education law, in relation to allowing pharmacy technicians to practice in any pharmacy under the supervision of a phar- macist PURPOSE: To remove the current limitation on the use of pharmacy technicians in pharmacies across New York, in order to support independent pharmacies. SUMMARY OF PROVISIONS: Section 1. Amends subdivision 1 of section 6841 of the education law, as added by chapter 414 of the laws of 2019. Section 2.