Bill S9359 version original - Jeremy Cooney
BILL NUMBER: S9359 SPONSOR: COONEY TITLE OF BILL: An act to amend the tax law, in relation to authorizing distributors of cannabis products to file annual returns electronically SUMMARY OF SPECIFIC PROVISIONS: Section 1: Subdivision (a) of section 495 of the tax law is amended. Section 2: Establishes the effective date. JUSTIFICATION: Under current law, the cannabis excise tax is paid quarterly by cannabis cultivators and processors to the NYS Department of Taxation and Finance.