Bill S9378 version original - Monica Martinez
BILL NUMBER: S9378 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the education law, in relation to establishing instruc- tion in financial and insurance literacy PURPOSE: The purpose of this legislation is to ensure all high school students are taught a curriculum that includes instruction in financial and insurance literacy. SUMMARY OF PROVISIONS: Section 1.