Bill S9392 version original - Sean Ryan
BILL NUMBER: S9392 SPONSOR: RYAN TITLE OF BILL: An act to amend chapter 731 of the laws of 2019 amending the executive law relating to regulatory fines for small businesses, in relation to the effectiveness thereof PURPOSE OR GENERAL IDEA OF BILL: Extends chapter 731 of the laws of 2019 relating to regulatory fines for small businesses for another three years. SUMMARY OF SPECIFIC PROVISIONS: Section 2 extends the expiration date of the existing law from December 20, 2024 to December 20, 2027.