Bill S8705A version A - Liz Krueger
BILL NUMBER: S8705A SPONSOR: KRUEGER TITLE OF BILL: An act to amend the domestic relations law, in relation to allowing for no fault separation PURPOSE OR GENERAL IDEA OF BILL: To amend the domestic relations law to add subdivision 6 to allow no fault separation. SUMMARY OF PROVISIONS: SECTION 1: Amends Subdivision 5 and 6 of section 170 of the domestic relations law to require a separation for a period of 6 months or more prior to the commencement of a no fault divorce.