Bill S9479 version original - Jeremy Cooney
BILL NUMBER: S9479 SPONSOR: COONEY TITLE OF BILL: An act to amend the transportation law and the executive law, in relation to work requiring members of the state police at certain work zones PURPOSE: This bill requires that a state police officer and car be present at certain work zones to increase safety. SUMMARY OF PROVISION: Section 1 defines this bill as DJ's law.