Bill S9432 version original - Lea Webb
BILL NUMBER: S9432 SPONSOR: WEBB TITLE OF BILL: An act to amend the tax law, in relation to authorizing the city of Binghamton to establish hotel and motel taxes PURPOSE: The purpose of this legislation is to authorize the city of Binghamton to establish an occupancy tax. SUMMARY OF PROVISIONS: Section 1: The tax law is amended by adding a new section 1202-jj. Section 2: Establishes the effective date.