Bill S9405 version original - Robert Rolison
BILL NUMBER: S9405 SPONSOR: ROLISON TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to authorizing the issuance of distinctive plates for medical technologists PURPOSE OR GENERAL IDEA OF BILL: This legislation authorizes the Department of Motor Vehicles to issue a distinctive license plate for medical technologists. SUMMARY OF PROVISIONS: Section 1: Adds new section 404-ff to the vehicle and traffic law to create distinctive plates for medical technologists.