Bill S9399 version original - James Skoufis
BILL NUMBER: S9399 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the penal law, in relation to adding possession of a flamethrower to the crime of criminal possession of a weapon in the fourth degree PURPOSE: The purpose of this bill is to establish the crime of criminal possession of a flamethrower SUMMARY OF PROVISIONS: Section 1. amends section 265.00 of the penal law by adding four new subdivisions 26, 27, 28, and 29 which provide for the definition of a flamethrower Section 2.