Bill S9424 version original - Brad Hoylman-Sigal
BILL NUMBER: S9424 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the family court act and the civil rights law, in relation to name changes of children 2. SOURCE OF BILL: This bill is being introduced at the request of the Judiciary, on the recommendation of the Chief Administrative Judge's Family Court Advisory and Rules Committee. 3.