Bill S9397 version original - William Weber
BILL NUMBER: S9397 SPONSOR: WEBER TITLE OF BILL: An act to amend the judiciary law, in relation to creating a fourteenth judicial district PURPOSE: To create a fourteenth judicial district. SUMMARY OF PROVISIONS: Section 1: amends section 70 of the judiciary law to account for the newly created fourteenth judicial district. Section 2: amends section 140 of the judiciary law to account for the fourteenth judicial district and amends the composition of the ninth judicial district.