Bill S9452 version original - Julia Salazar
BILL NUMBER: S9452 Revised 6/3/24 SPONSOR: SALAZAR SOURCE OF BILL: This bill is being introduced at the request of the Judiciary, upon the recommendation of the Chief Administrative Judge's Advisory Committee on Civil Practice. PURPOSE OF BILL: This bill eliminates the outdated and pejorative designation of "poor person" status for court users who lack sufficient means to pay the costs, fees, and expenses necessary to prosecute or defend an action or appeal.