Bill S9468 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S9468 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to licensure fees for seniors SUMMARY OF PROVISIONS: Section 1 amends paragraph c of subdivision 2 of section 503 of the vehicle and traffic law to waive the fee of license renewal for those who are sixty-two years of age or older. Section 2 sets the effective date. JUSTIFICATION: The senior population in New York State is larger and living longer than ever before.