Bill S9440 version original - Jeremy Cooney
BILL NUMBER: S9440 SPONSOR: COONEY TITLE OF BILL: An act to repeal article 21 of the tax law relating to the highway use tax PURPOSE: This bill repeals the highway use tax. SUMMARY OF PROVISIONS:ò Repeals article 21 of the tax law relating to imposition of a highway use tax for the privilege of operating any vehicular unit upon the public high- ways of the state.