Bill S9447 version original - George Borrello
BILL NUMBER: S9447 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the real property actions and proceedings law, in relation to the sale of a mortgaged premises after a judgment of sale has been entered PURPOSE OR GENERAL IDEA OF BILL: This bill relates to the sale of a mortgaged premises after a judgment of sale has been entered. SUMMARY OF PROVISIONS: Section 1.