Bill S9474 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S9474 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to providing free use of state parks for members or citizens of state or federally recognized Native American nations. PURPOSE:; This legislation amends the parks, recreation, and historic preservation law to allow any member or citizen of a state or federally recognized Native American Nation to have free access to state parks and campsites.