Bill S9442 version original - Jeremy Cooney
BILL NUMBER: S9442 SPONSOR: COONEY TITLE OF BILL: An act to amend the cannabis law, in relation to adult-use cultivators PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to enact legislation that allows for Adult- Use Conditional Cultivators to select a permanent license type that allows them to meet the demand of a growing state cannabis market. SUMMARY OF PROVISIONS: Section 1 provides the short title "cannabis adult-use cultivator fair- ness act.