Bill S8553C version C - Joseph P. Addabbo Jr.
BILL NUMBER: S8553C SPONSOR: ADDABBO TITLE OF BILL: An act to amend the insurance law and the social services law, in relation to mandatory health insurance coverage for follow-up screening or diagnostic services for lung cancer PURPOSE: Lung cancer is one of the most common cancers among New Yorkers. In New York it is the leading cause of cancer deaths. Each year over 6,700 men and over 7,200 women are diagnosed with lung cancer and about 3,800 men and over 3,600 women die from this disease.