Bill S9398 version original - Jessica Ramos
BILL NUMBER: S9398 SPONSOR: RAMOS TITLE OF BILL: An act to amend the workers' compensation law, the labor law, and the executive law, in relation to requiring the workers' compensation board and the department of labor to publish lists of debarred entities as open data PURPOSE: To increase transparency of the list of entities that have been debarred for labor and other violations through publication of lists as open data. SUMMARY OF PROVISIONS: Section 1.