Bill S8379A version A - Jessica Ramos
BILL NUMBER: S8379A SPONSOR: RAMOS TITLE OF BILL: An act in relation to establishing the task force to promote the employ- ment of people with disabilities PURPOSE: This bill establishes a task force to promote competitive integrated employment for people with disabilities. SUMMARY OF PROVISIONS: Section 1 establishes a task force in the Department of Labor to promote competitive integrated employment for people with disabilities.