Bill S9477 version original - Jeremy Cooney
BILL NUMBER: S9477 SPONSOR: COONEY TITLE OF BILL: An act to amend the banking law, in relation to establishing joint and survivorship accounts SUMMARY OF PROVISIONS: Section 1. Amends the banking law by adding section 675-a, which deline- ates the parameters of joint accounts, joint and survivorship accounts, regulations of the superintendent, and the bill's application.