Bill S9441 version original - Jeremy Cooney
BILL NUMBER: S9441 SPONSOR: COONEY TITLE OF BILL: An act to amend the state finance law and the cannabis law, in relation to authorizing a qui tam proceeding for civil violations of the cannabis law PURPOSE OR GENERAL IDEA OF BILL: This bill would permit a person, known as a qui tam plaintiff, to bring a civil action under the New York False Claims Act for violations of general prohibitions and restrictions of cannabis law. SUMMARY OF PROVISIONS: Section 1.