Bill S9444 version original - Sean Ryan
BILL NUMBER: S9444 SPONSOR: RYAN TITLE OF BILL: An act to amend the judiciary law, in relation to the creation of the fourteenth judicial district consisting of the county of Erie and a supreme court library at Erie PURPOSE: This bill would create a new judicial district made up of the county of Erie. SUMMARY OF PROVISIONS: Section 1. Amends section 70 of the judiciary law to add a reference to the fourteenth judicial district.