Bill S9463 version original - Dean Murray
BILL NUMBER: S9463 REVISED 05/16/2024 SPONSOR: MURRAY TITLE OF BILL: An act to amend the public health law, in relation to enacting "Josh's law", requires training plans be approved prior to implementation in modified and high school football PURPOSE: This bill enacts "Josh's Law" to require that training plans be approved prior to implementation in modified and high school football. SUMMARY OF PROVISIONS: Section I states that this act shall be known and may be cited as "Josh's Law.