Bill S9425B version original - William Weber
BILL NUMBER: S9425 SPONSOR: WEBER TITLE OF BILL: An act authorizing Kolel Rachmistrivka to receive retroactive real prop- erty tax exempt status PURPOSE: To authorize the assessor of the Town of Ramapo to accept an application for exemption from real property taxes with respect to the 2024 - 2025 school taxes and all of the 2025 general taxes for. Kolel Rachmistriv- ka. SUMMARY OF PROVISIONS: Section 1.