Bill S9644 version original - Iwen Chu
BILL NUMBER: S9644 SPONSOR: CHU TITLE OF BILL: An act to amend the education law, in relation to establishing a course of study in financial education for certain students PURPOSE: To establish a course of study in financial education for certain students SUMMARY OF PROVISIONS: Section 1. Adds a new section 803-b to the education law Section 2. Establishes the effective date JUSTIFICATION: Teenagers and young adults are increasingly responsible for their own finances.