Bill S9651 version original - Samra Brouk
BILL NUMBER: S9651 SPONSOR: BROUK TITLE OF BILL: An act to amend the insurance law, in relation to expanding the existing definition of wellness programs PURPOSE OR GENERAL IDEA OF BILL: To establish insurance coverage for wellness program(s) to address impacts of technology use on the mental and physical well being of chil- dren, adolescents, and young adults SUMMARY OF PROVISIONS: Section 1. Legislative Findings. Section 2.