Bill S9624 version original - Jeremy Cooney
BILL NUMBER: S9624 SPONSOR: COONEY TITLE OF BILL: An act to amend the labor law, in relation to a minimum operator age for amusement devices SUMMARY OF PROVISIONS: Section 1. The labor is amended by adding a new section 870-p to require the Department of Labor to promulgate regulations to lower the minimum age required to operate all amusement park devices. Section 2. is the effective date.