Bill S9648 version original - Iwen Chu
BILL NUMBER: S9648 SPONSOR: CHU TITLE OF BILL: An act to amend the education law, in relation to a registry for histor- ic libraries and museums PURPOSE: The purpose of this legislation is to identify historically significant libraries and museums throughout New York State. SUMMARY OF PROVISIONS: Section 1 directs the Department of Education to establish and maintain a registry of historic libraries and museums in the state.