Bill S9605 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S9605 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act in relation to authorizing the county of Nassau assessor to accept an application for a retroactive real property tax exemption from the Center for Jewish Life of Hewlett Inc. PURPOSE: An act in relation to authorizing the county of Nassau assessor to accept an application for a retroactive real property tax exemption from the Center for Jewish Life of Hewlett Inc. SUMMARY OF PROVISIONS: Section 1.