Bill S8402A version A - Kevin S. Parker
BILL NUMBER: S8402A SPONSOR: PARKER TITLE OF BILL: An act to amend the public authorities law, in relation to establishing the "go green schools act" to assist schools in converting to renewable energy systems as a primary source of energy PURPOSE OR GENERAL IDEA OF BILL: AN ACT that will enable and incentivize the transition of schools in the state to become renewable energy operated as a primary source of energy usage. SUMMARY OF PROVISIONS: 1. Provides Short title of the Bill 2.