Bill S9630 version original - Monica Martinez
BILL NUMBER: S9630 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the public health law, in relation to limited-scope radiographers PURPOSE: To create a registration and licensing process for limited-scope radiog- raphers. SUMMARY OF PROVISIONS: Section 1: Allows the department to issue separate licenses in the areas of radiography, radiation therapy, or nuclear medicine technology to individuals who qualify under section thirty-five of this title.