Bill S9603 version original - Steve Rhoads
BILL NUMBER: S9603 SPONSOR: RHOADS TITLE OF BILL: An act to amend the criminal procedure law, in relation to establishing that making a terroristic threat is a qualifying offense for bail PURPOSE: Establish that making a terroristic threat as a qualifying offense for bail. SUMMARY OF PROVISIONS: Section 1. Paragraph (g) of subdivision 4 of section 510.