Bill S9498 version original - Jeremy Cooney
BILL NUMBER: S9498 SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to dealer sale rescission or cancelation for any vehicle PURPOSE: To permit the cancellation or rescinding of the sale of a vehicle with approval by both dealer and purchaser. Section 1 adds a new section 2124-a to the vehicle and traffic law Section 2 sets forth the effective date.