Bill S9502 version original - Jeremy Cooney
BILL NUMBER: S9502 SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to eligibility for participation in the pre-licensing course internet pilot program PURPOSE: To remove bathers to online instruction for the five-hour pre-licensing course. SUMMARY OF PROVISIONS: Section one of this bill removes a prohibition on drivers holding class DJ and MJ Licenses from participating in online pre-licensing courses. Section two provides the effective date.