Bill S9501 version original - Jeremy Cooney
BILL NUMBER: S9501 SPONSOR: COONEY TITLE OF BILL: An act to amend the highway law, in relation to authorizing and direct- ing the department of transportation to establish a memorial marker program PURPOSE: To establish a formal program in New York to allow for the placement of memorials at the location of fatal highway accidents on state highways and roads. Section 1 adds a new section 67 to highway law. Section 2 establishes the effective date. JUSTIFICATION:.