Bill S9529 version original - Jessica Scarcella-Spanton
BILL NUMBER: S9529 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the veterans' services law, in relation to the employ- ment of veterans with disabilities SUMMARY OF PROVISIONS: Section 1. Amends subdivision 13 of section 4 of the veterans' services law by adding the New York State Department of Civil Service to the list of agencies who must maintain a veteran's employment portal on their website. Section 2.