Bill S9657 version original - Monica Martinez
BILL NUMBER: S9657 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the trans- portation of dogs PURPOSE: The purpose of this legislation is to prohibit the transportation of dogs in the back of a motorvehicle in a space intended for any load with exceptions. SUMMARY OF PROVISIONS: Section 1.