Bill S9619 version original - Jeremy Cooney
BILL NUMBER: S9619 SPONSOR: COONEY TITLE OF BILL: An act to amend the penal law, in relation to assaults on medical providers and employees and volunteers of hospitals PURPOSE: To include physicians, nurses, ambulance, directors, employees, agents and volunteers of hospitals performing an assigned duty within the Class D felony of assault in the second degree. SUMMARY OF PROVISIONS: Section one amends subdivision 3 and 11 of section 120.