Bill S9492 version original - Jeremy Cooney
BILL NUMBER: S9492 SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to providing for the electronic transmission of certain documents PURPOSE OR GENERAL IDEA OF BILL: To allow for the electronic transmission of vehicle titles from the Department of Motor Vehicles to the vehicle owner. SUMMARY OR SPECIFIC PROVISIONS: Section I amends Vehicle and Traffic Law Article 46 Section 2101 to include electronic transmission.