Bill S9572 version original - Cordell Cleare
BILL NUMBER: S9572 SPONSOR: CLEARE TITLE OF BILL: An act requiring the empire state development corporation to establish an office of accountability and compliance PURPOSE: To require ESDC to take a more active role in monitoring the projects it approves for accountability and compliance. SUMMARY OF PROVISIONS: This bill requires ESDC to take a more active role in monitoring the projects it approves for accountability and compliance.