Bill S9667 version original - Jamaal Bailey
BILL NUMBER: S9667 SPONSOR: BAILEY TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "Dwight 'Heavy D' Myers Way" PURPOSE: This bill renames a portion of the state highway system in the City of Mount Vernon after "Dwight 'Heavy D' Myers". SUMMARY OF PROVISIONS: Section 1: The highway law is amended by adding a new section 344-p § 2: Requires the commissioner of transportation to provide installation and maintenance of signage.