Bill S9602 version original - Steve Rhoads
BILL NUMBER: S9602 SPONSOR: RHOADS TITLE OF BILL: An act to amend the public health law, in relation to requiring all medical facilities to be able to access every single-occupancy bathroom where a patient could be in an emergency PURPOSE: Requires all medical facilities to have immediate access to every single-occupancy bathroom where a patient could be in an emergency. SUMMARY OF PROVISIONS: Section 1. The public health law is amended by adding a new section 2831 to read as follows: § 2831.