Bill S9601 version original - Steve Rhoads
BILL NUMBER: S9601 SPONSOR: RHOADS TITLE OF BILL: An act to amend the judiciary law, in relation to recusal by a court of appeals judge PURPOSE: Relates to recusal by a court of appeals judge. SUMMARY OF PROVISIONS: Section 1. Section 9 of the judiciary law, as added by chapter 376 of the laws of 2020, is amended to read as follows: § 9. Recusal; reason. 1.