Bill S9510 version original - Jeremy Cooney
BILL NUMBER: S9510 SPONSOR: COONEY TITLE OF BILL: An act to amend the highway law, in relation to indexing appropriations PURPOSE AND JUSTIFICATION: Cost increases are inevitable and do not discriminate against individ- uals or industries. This reality is particularly true in the highway construction industry, where the price of materials and the fuel neces- sary to deliver them have skyrocketed in recent years.