Bill S9623 version original - Jeremy Cooney
BILL NUMBER: S9623 SPONSOR: COONEY TITLE OF BILL: An act to amend the state finance law, in relation to the discretionary buying thresholds SUMMARY OF SPECIFIC PROVISIONS: Section one amends subdivision 6 of section 163 of the state finance law to increase the discretionary bidding thresholds for contracts with small businesses and MWBEs to $1 million. Section two establishes an effective date.