Bill S9573 version original - Lea Webb
BILL NUMBER: S9573 SPONSOR: WEBB TITLE OF BILL: An act to amend the tax law, in relation to authorizing the town of Dickinson to establish hotel and motel taxes; and providing for the repeal of such provisions upon expiration thereof PURPOSE: The purpose of this legislation is to authorize the town of Dickinson to establish an occupancy tax. SUMMARY OF PROVISIONS: Section 1: The tax law is amended by adding a new section 1202-jj. Section 2: Establishes the effective date.