Bill S9503 version original - Jeremy Cooney
BILL NUMBER: S9503 SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to creating the "4-H member" distinctive license plate; and to amend the state finance law, in relation to establishing the 4-H foundation fund PURPOSE: The purpose of this bill is to provide any person residing in this state, upon request, a distinctive "4 -H member" license plate. SUMMARY OF PROVISIONS: Section 1. The vehicle and traffic law is amended by adding a new section 404-y.