Bill S9557 version original - Jeremy Cooney
BILL NUMBER: S9557 SPONSOR: COONEY TITLE OF BILL: An act to amend the cannabis law, in relation to rules and regulations for advertising and marketing of cannabis SUMMARY OF PROVISIONS: Section 1. Amends section 86 of the cannabis law to require the board to promulgate rules and regulations for advertising and marketing allowing for certain signage, window tinting, lighting, loyalty programs, and distribution of flyers. Section 2. Sets the effective date.