Bill S9520 version original - Shelley Mayer
BILL NUMBER: S9520 SPONSOR: MAYER TITLE OF BILL: An act to amend the local finance law, in relation to the exclusion of certain indebtedness of school districts in a city with one hundred twenty-five thousand inhabitants or more PURPOSE OR GENERAL IDEA OF BILL: This bill changes the way debt is calculated for small city school districts to align with other fiscally independent districts. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 121.