Bill S9673A version original - Joseph P. Addabbo Jr.
BILL NUMBER: S9673 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law, in relation to the acceleration of the downstate casino licenses PURPOSE OR GENERAL IDEA OF BILL: To accelerate the downstate casino licensing review and award process. SUMMARY OF PROVISIONS: Section 1 of this bill amends paragraph 7 of section 109-a of the racing, pari-mutuel wagering and breeding law.