Bill S9149A version A - Nathalia Fernandez
BILL NUMBER: S9149A SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the public health law, in relation to prohibiting charg- es for the issuance of a certificate of still birth or fetal death PURPOSE OR GENERAL IDEA OF BILL: This bill prohibits fees associated with obtaining a certificate of stillbirth or fetal death. SUMMARY OF PROVISIONS: Section 1.