Bill S9674 version original - Michael Gianaris
BILL NUMBER: S9674 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the judiciary law and the criminal procedure law, in relation to the publication of opinions rendered in criminal causes PURPOSE: To increase transparency of New York legal decisions in criminal proceedings. SUMMARY OF PROVISIONS: Section 1 of the bill provides that the short title of the act is the Criminal Court Opinion Transparency Act.