Bill S9678B version original - Kristen Gonzalez
BILL NUMBER: S9678 SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the election law, in relation to materially deceptive media in political communications PURPOSE: This bill would amend the election law, in relation to materially decep- tive media in political communications. SUMMARY OF PROVISIONS: Section 1. Amends paragraph (b) of subsection 5 of section 12-106 of the election law as follows.