Bill S4459A version A - Jake Ashby
BILL NUMBER: S4459A SPONSOR: ASHBY TITLE OF BILL: An act to amend the state law, in relation to designating the state of New York a purple heart state PURPOSE: To amend the state law, in relation to designating the state of New York a Purple Heart state. SUMMARY OF PROVISIONS: Section 1. State law is amended by adding section 92. This designates the state of New York as a Purple Heart State Section 2 sets forth the effective date.