Bill S5500B version B - James Skoufis
BILL NUMBER: S5500B SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the public officers law, in relation to the disclosure of public employee disciplinary records PURPOSE: To provide all public employees with notice whenever their personal information has been released to the public pursuant to a FOIL request. SUMMARY OF PROVISIONS: Section 1 of the bill amends section 87 of the Public Officers Law by adding a new subdivision 4-c to effectuate the purposes of the legis- lation.