Bill S2078B version B - Brian Kavanagh
BILL NUMBER: S2078B SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the multiple dwelling law and the multiple residence law, in relation to the use of smart access systems and the information that may be gathered from such systems PURPOSE: This legislation relates to limitations on electronic or computerized systems for entry. SUMMARY OF PROVISIONS: Section 1 of the bill amends the multiple dwelling law by adding a new section 50-b.