Bill S7807A version A - José M. Serrano
BILL NUMBER: S7807A SPONSOR: SERRANO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to establishing standards for trail closures on the Empire state trail PURPOSE: To direct the office of parks, recreation and historical preservation to coordinate with an owner of a portions of the Empire State Trail that is temporarily closed to establish a signed detour and to post notice on the office's website. SUMMARY OF PROVISIONS: Section 1.