Bill S5703B version A - Kevin S. Parker
BILL NUMBER: S5703A SPONSOR: PARKER TITLE OF BILL: An act to amend the general business law, in relation to prohibiting the use of social media platforms, email or text messages for the purposes of collecting debts PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to prevent debt collectors from using online contact information to collect on a consumer debt. SUMMARY OF PROVISIONS: Section 1: Section 601 of the general business law is amended by adding two new subdivisions 12 and 13.