Bill S1295A version A - James Sanders Jr.
BILL NUMBER: S1295A SPONSOR: SANDERS TITLE OF BILL: An act to amend the civil practice law and rules, in relation to time limitations for filing claims for certain injuries PURPOSE OR GENERAL IDEA OF BILL: To amend the civil practice law and rules in relation to time limita- tions for filing claims for certain injuries. SUMMARY OF SPECIFIC PROVISIONS: Subdivision 4 of section 214-c of the civil practice laws and rules, as added by chapter 682 of the laws of 1986 is amended.