Bill S9696 version original - Jabari Brisport
BILL NUMBER: S9696 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the family court act, in relation to the right to coun- sel in proceedings regarding violations of orders of child support and to establish paternity or parentage in the family court 2. SOURCE OF BILL This bill is being introduced at the request of the Judiciary, upon the recommendation of the Chief Administrative Judge's Family Court Advisory and Rules Committee. 3.