Bill S9708 version original - Kevin Thomas
BILL NUMBER: S9708 SPONSOR: THOMAS TITLE OF BILL: An act to amend the penal law, in relation to deterring theft of essen- tial infrastructure PURPOSE: To update the existing offense of grand larceny penalties when an indi- vidual knowingly damages or tampers with critical infrastructure intend- ing to disrupt utility services, including wireless and broadband services. SUMMARY OF PROVISIONS: Section 1 names the act the "Prevention of Damage to Essential Infras- tructure Act of 2024.