Bill S4196A version A - James Sanders Jr.
BILL NUMBER: S4196A SPONSOR: SANDERS TITLE OF BILL: An act to amend the penal law, in relation to the definition of consent PURPOSE OR GENERAL IDEA OF BILL:: Click here to enter text. SUMMARY OF PROVISIONS:: Section 1 of this legislation defines consent as a "Freely Given, Know- ledgeable and Informed Agreement." DIFFERENCE BETWEEN ORIGINAL AND AMENDED VERSION (IF APPLICABLE): Click here to enter text.