Bill S9705 version original - Jamaal Bailey
BILL NUMBER: S9705 SPONSOR: BAILEY TITLE OF BILL: An act to amend the tax law, in relation to establishing a MWBE or SDVOBE subcontractor tax credit PURPOSE: This act will amend the tax law to include a MWBE or SDVOBE subcontrac- tor tax incentive. SUMMARY OF PROVISIONS: Section 210-B of the tax law is amended by adding a new subdivision 60 which establishes the MWBE or SDVOBE subcontractor tax credit.