Bill S9271A version A - Michelle Hinchey
BILL NUMBER: S9271A SPONSOR: HINCHEY TITLE OF BILL: An act to amend the tax law, in relation to establishing an occupancy tax in the village of Coxsackie; and providing for the repeal of such provisions upon expiration thereof PURPOSE:: To authorize the village of Coxsackie to establish an occupancy tax. SUMMARY OF PROVISIONS:: Section I: Authorizes the village of Coxsackie in Greene County to adopt local laws imposing an occupancy tax at a rate of 4% of the per diem rental rate for each room.