Bill S9130A version A - Sean Ryan
BILL NUMBER: S9130A SPONSOR: RYAN TITLE OF BILL: An act to amend the state finance law, in relation to the use of funds and payments from the IOLA fund PURPOSE: To ensure that monies from the IOLA fund are used only for the fund's original, statutorily intended purpose. SUMMARY OF PROVISIONS: Section 1: amends paragraph f of subdivision 3 of section 97-v of the state finance law by prohibiting monies in the IOLA fund to be used for any purpose other than the provision of legal services.