Bill S9628A version A - James Skoufis
BILL NUMBER: S9628A SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the retirement and social security law, in relation to providing accidental disability retirement for deputy sheriffs in Orange county PURPOSE: To provide the option for an accidental disability retirement benefit of 75% of the final average salary for deputy sheriffs employed by Orange County. SUMMARY OF PROVISIONS: Section 1 adds a new Section 605-g to the Retirement and Social Security Law.