Bill S9736 version original - Luis R. Sepúlveda
BILL NUMBER: S9736 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend chapter 890 of the laws of 1982, relating to the estab- lishment of certain water charges for hospitals and charities in New York city, in relation to the effectiveness thereof PURPOSE OR GENERAL IDEA OF BILL: This legislation limits water charges for hospitals and charitable organizations. SUMMARY OF PROVISIONS: This bill amends Chapter 890 of the Laws of 1982 to extend its expira- tion date until September 1, 2026.